requirements.txt
numpy>=1.24
pandas>=2.0
scipy>=1.10
PyYAML>=6.0
scikit-learn>=1.3
tqdm>=4.66
torch>=2.2
pytest>=8.0
